Move logic complete
