+++ src/logic.js
@@ -1,12 +1,22 @@
-const listCountElement = document.querySelector('[data-list-count]');
 const listsContainer = document.querySelector('[data-lists]');
+const newListForm = document.querySelector('[data-new-list-form]');
+const newListinput = document.querySelector('[data-new-list-input]');
+const deleteListButton = document.querySelector('[data-delete-list-button]');
+const listDisplayContainer = document.querySelector('[data-list-display-container]');
+const listTitleElement = document.querySelector('[data-list-title]');
+const listCountElement = document.querySelector('[data-list-count]');
+const tasksContainer = document.querySelector('[data-tasks]');
+const taskTemplate = document.getElementById('task-template');
+const newTaskForm = document.querySelector('[data-new-task-form]');
+const newTaskInput = document.querySelector('[data-new-task-input]');
+const clearCompleteTasksButton = document.querySelector('[data-clear-complete-tasks-button]');
 
-  
-  const LOCAL_STORAGE_LIST_KEY = 'task.lists';
-  const LOCAL_STORAGE_SELECTED_LIST_ID_KEY = 'task.selectedListsId';
+const LOCAL_STORAGE_LIST_KEY = 'task.lists';
+const LOCAL_STORAGE_SELECTED_LIST_ID_KEY = 'task.selectedListsId';
 
-  let lists = JSON.parse(localStorage.getItem(LOCAL_STORAGE_LIST_KEY)) || [];
-  let selectedListId = localStorage.getItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY);
+let lists = JSON.parse(localStorage.getItem(LOCAL_STORAGE_LIST_KEY)) || [];
+let selectedListId = localStorage.getItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY);
+
   function createList(name) {
     return { id: Date.now().toString(), name, tasks: [] };
   }
@@ -21,29 +31,12 @@ const listsContainer = document.querySelector('[data-lists]');
     }
   }
 
-  function save() {
-    localStorage.setItem(LOCAL_STORAGE_LIST_KEY, JSON.stringify(lists));
-    localStorage.setItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY, selectedListId);
-  }
 
   function renderTaskCount(selectedListId) {
     const incompleteTaskCount = selectedListId.tasks.filter((task) => !task.complete).length;
     const taskString = incompleteTaskCount === 1 ? 'task' : 'tasks';
     listCountElement.innerText = `${incompleteTaskCount} ${taskString} remaining`;
   }
+  
 
-  function renderLists() {
-    lists.forEach((list) => {
-      const listElement = document.createElement('li');
-      listElement.dataset.listId = list.id;
-      listElement.classList.add('list-name');
-      listElement.innerText = list.name;
-      if (list.id === selectedListId) {
-        listElement.classList.add('active-list');
-      }
-      listsContainer.appendChild(listElement);
-    });
-  }
-
-
-  export { renderLists, createList, createTask, clearElement, save, renderTaskCount}
+  export { createList, createTask, clearElement, renderTaskCount}
